src/main/java/com/mostafayehya/deliciousrecipes/converters/CategoryCommandToCategory.java
package com.mostafayehya.deliciousrecipes.converters;

import com.mostafayehya.deliciousrecipes.comands.CategoryCommand;
import com.mostafayehya.deliciousrecipes.domain.Category;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class CategoryCommandToCategory implements Converter<CategoryCommand, Category> {

    @Synchronized
    @Nullable
    @Override
    public Category convert(CategoryCommand source) {

        if (source == null)
            return null;

        Category category = new Category();

        category.setId(source.getId());
        category.setDescription(source.getDescription());
        return category;

    }
}
